requirements.txt
chia-blockchain==1.3.1
fastapi==0.73.0
async-timeout<4.0,>=3.0
 
uvicorn==0.17.0.post1
websockets~=8.1.0
aioredis==1.3.0
logzero
aiocache